setup useEffect module for auth
