+++ src/app/mypage/page.tsx
@@ -1,49 +1,20 @@
 "use client";
 
-import { useRouter } from "next/navigation";
-import { useState, useEffect } from "react";
+import { useState } from "react";
 
 import GroupSection from "./GroupSection";
 import ProfileSection from "./ProfileSection";
 
 import { LoadingCircle } from "@/components/LoadingCircle";
-import { fetchUserInfo } from "@/lib/fetch";
-import { getAccessToken, useUserContext } from "@/lib/getAccessToken";
-import { UserInfo } from "@/lib/schema";
+import authEffect from "@/lib/authEffect";
+import { getAccessToken, useUserContext } from "@/lib/dataUtils";
 
 const MyPage = () => {
 	const [userInfo, setUserInfo] = useUserContext();
-	const [loading, setLoading] = useState<boolean>(false);
+	const [loading, setLoading] = useState<boolean>(true);
 	const accessToken = getAccessToken();
 
-	const router = useRouter();
-
-	useEffect(() => {
-		const fetchUserData = async () => {
-			setLoading(true);
-			try {
-				const response = await fetchUserInfo(accessToken);
-
-				if (response.status === 200) {
-					const newUserInfo: UserInfo = await response.json();
-					setUserInfo(newUserInfo);
-					return;
-				} else {
-					const { error } = await response.json();
-					setUserInfo(null);
-					router.push("/");
-					console.error(error);
-					return;
-				}
-			} finally {
-				setLoading(false);
-			}
-		};
-
-		if (!userInfo) {
-			fetchUserData();
-		}
-	}, [accessToken]);
+	authEffect(accessToken, setLoading, userInfo, setUserInfo);
 
 	if (loading) {
 		return (
